rename: CustomStringUtils로 클래스 이름 변경
- org.springframework.util.StringUtils와 겹쳐서 클래스 이름 변경
- 토큰만 추출하는 메서드 추가

#19
