+++ user-service/src/main/java/com/pda/user_service/dto/KakaoUserDto.java
@@ -3,7 +3,7 @@ package com.pda.user_service.dto;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 import com.pda.user_service.jpa.User;
-import com.pda.utils.api_utils.StringUtils;
+import com.pda.utils.api_utils.CustomStringUtils;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -48,7 +48,7 @@ public class KakaoUserDto {
 
     public User toEntity(String token) {
         String birthdate = kakaoAccount.birthyear + kakaoAccount.birthday;
-        String phone = StringUtils.formatPhone(kakaoAccount.phoneNumber);
+        String phone = CustomStringUtils.formatPhone(kakaoAccount.phoneNumber);
         return new User(kakaoAccount.email, kakaoAccount.name, token, phone, birthdate, kakaoAccount.gender);
     }
 }
